+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Remote Procedure Calls for Pip.Services in Dart Changelog
+# <img src="https://uploads-ssl.webflow.com/5ea5d3315186cf5ec60c3ee4/5edf1c94ce4c859f2b188094_logo.svg" alt="Pip.Services Logo" width="200"> <br/> Remote Procedure Calls for Dart Changelog
 
 ## 1.0.0